Create components for displaying emails
