+++ src/App.jsx
@@ -1,5 +1,7 @@
 import { useState } from 'react'
 
+import Emails from './Components/Emails'
+
 import initialEmails from './data/emails'
 
 import './styles/App.css'
@@ -16,23 +18,23 @@ function App() {
   const unreadEmails = emails.filter(email => !email.read)
   const starredEmails = emails.filter(email => email.starred)
 
-  const toggleStar = targetEmail => {
-    const updatedEmails = emails =>
-      emails.map(email =>
-        email.id === targetEmail.id
-          ? { ...email, starred: !email.starred }
-          : email
-      )
-    setEmails(updatedEmails)
-  }
+  // const toggleStar = targetEmail => {
+  //   const updatedEmails = emails =>
+  //     emails.map(email =>
+  //       email.id === targetEmail.id
+  //         ? { ...email, starred: !email.starred }
+  //         : email
+  //     )
+  //   setEmails(updatedEmails)
+  // }
 
-  const toggleRead = targetEmail => {
-    const updatedEmails = emails =>
-      emails.map(email =>
-        email.id === targetEmail.id ? { ...email, read: !email.read } : email
-      )
-    setEmails(updatedEmails)
-  }
+  // const toggleRead = targetEmail => {
+  //   const updatedEmails = emails =>
+  //     emails.map(email =>
+  //       email.id === targetEmail.id ? { ...email, read: !email.read } : email
+  //     )
+  //   setEmails(updatedEmails)
+  // }
 
   let filteredEmails = emails
 
@@ -87,35 +89,8 @@ function App() {
           </li>
         </ul>
       </nav>
-      <main className="emails">
-        <ul>
-          {filteredEmails.map((email, index) => (
-            <li
-              key={index}
-              className={`email ${email.read ? 'read' : 'unread'}`}
-            >
-              <div className="select">
-                <input
-                  className="select-checkbox"
-                  type="checkbox"
-                  checked={email.read}
-                  onChange={() => toggleRead(email)}
-                />
-              </div>
-              <div className="star">
-                <input
-                  className="star-checkbox"
-                  type="checkbox"
-                  checked={email.starred}
-                  onChange={() => toggleStar(email)}
-                />
-              </div>
-              <div className="sender">{email.sender}</div>
-              <div className="title">{email.title}</div>
-            </li>
-          ))}
-        </ul>
-      </main>
+
+      <Emails filteredEmails={filteredEmails} setEmails={setEmails}/>
     </div>
   )
 }
