+++ src/Components/Emails.jsx
@@ -1,0 +1,13 @@
+import Email from "./Email";
+
+export default function Emails({filteredEmails, setEmails}) {
+    return (
+        <main className='Emails'>
+            <ul>
+                {filteredEmails.map((email, index) => (
+                    <Email key={index} email={email} setEmails={setEmails} />
+                ))}
+            </ul>
+        </main>
+    );
+}
