+++ src/data/emails.js
@@ -1,37 +1,47 @@
 export default [
-  {
-    id: 1,
-    sender: `Zoom`,
-    title: `Cloud Recording - Nicolas Marcora's Personal Meeting Room is now available`,
-    starred: false,
-    read: true
-  },
-  {
-    id: 2,
-    sender: `Zoom`,
-    title: `Sean Davison has joined your Personal Meeting Room`,
-    starred: false,
-    read: false
-  },
-  {
-    id: 3,
-    sender: `Notion`,
-    title: `1 update in Boolean`,
-    starred: true,
-    read: true
-  },
-  {
-    id: 4,
-    sender: `The Calendly Team`,
-    title: `Use more than one calendar?`,
-    starred: false,
-    read: false
-  },
-  {
-    id: 5,
-    sender: `Patrick`,
-    title: `Updated invitation: Coding chat with Nico`,
-    starred: true,
-    read: false
-  }
+	{
+		id: 1,
+		sender: `Zoom`,
+		title: `Cloud Recording - Nicolas Marcora's Personal Meeting Room is now available`,
+		starred: false,
+		read: true,
+		mailTxt:
+			"Lorem ipsum dolor, sit amet consectetur adipisicing elit. Totam, molestias distinctio ipsa debitis tempora nihil obcaecati, a consequatur doloribus, iste excepturi eius autem quae odio accusantium eum pariatur consequuntur aperiam?",
+	},
+	{
+		id: 2,
+		sender: `Zoom`,
+		title: `Sean Davison has joined your Personal Meeting Room`,
+		starred: false,
+		read: false,
+		mailTxt:
+			"Lorem ipsum dolor, sit amet consectetur adipisicing elit. Totam, molestias distinctio ipsa debitis tempora nihil obcaecati, a consequatur doloribus, iste excepturi eius autem quae odio accusantium eum pariatur consequuntur aperiam?",
+	},
+	{
+		id: 3,
+		sender: `Notion`,
+		title: `1 update in Boolean`,
+		starred: true,
+		read: true,
+		mailTxt:
+			"Lorem ipsum dolor, sit amet consectetur adipisicing elit. Totam, molestias distinctio ipsa debitis tempora nihil obcaecati, a consequatur doloribus, iste excepturi eius autem quae odio accusantium eum pariatur consequuntur aperiam?",
+	},
+	{
+		id: 4,
+		sender: `The Calendly Team`,
+		title: `Use more than one calendar?`,
+		starred: false,
+		read: false,
+		mailTxt:
+			"Lorem ipsum dolor, sit amet consectetur adipisicing elit. Totam, molestias distinctio ipsa debitis tempora nihil obcaecati, a consequatur doloribus, iste excepturi eius autem quae odio accusantium eum pariatur consequuntur aperiam?",
+	},
+	{
+		id: 5,
+		sender: `Patrick`,
+		title: `Updated invitation: Coding chat with Nico`,
+		starred: true,
+		read: false,
+		mailTxt:
+			"Lorem ipsum dolor, sit amet consectetur adipisicing elit. Totam, molestias distinctio ipsa debitis tempora nihil obcaecati, a consequatur doloribus, iste excepturi eius autem quae odio accusantium eum pariatur consequuntur aperiam?",
+	},
 ]
