+++ src/styles/App.css
@@ -1,6 +1,6 @@
 .app {
   height: 100vh;
-
+  background-color: darkslategray;
   display: grid;
   grid-template:
     'header header' 50px
